backend/app/services/gmail.py
import os.path
from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build
import logging

logger = logging.getLogger(__name__)

SCOPES = [
    'https://www.googleapis.com/auth/gmail.readonly',
    'https://www.googleapis.com/auth/gmail.send',
    'https://www.googleapis.com/auth/calendar.readonly'
]

def get_gmail_service():
    # Placeholder: In production, we'd load credentials from settings or volume mount
    # For now, return None and log caution
    logger.warning("Gmail Service: Credentials not yet configured.")
    return None

async def list_recent_emails(service, max_results=10):
    if not service: return []
    results = service.users().messages().list(userId='me', maxResults=max_results).execute()
    return results.get('messages', [])

async def send_email(service, to, subject, body):
    if not service: return False
    # Implementation for sending email
    return True
