backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
redis
chromadb
openai
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-multipart
celery
prometheus-fastapi-instrumentator
